+++ keeper/src/test/scala/zio/keeper/ByteCodecSpec.scala
@@ -1,15 +1,15 @@
 package zio.keeper
 
-import zio.keeper.membership.hyparview.ActiveProtocol
-import zio.keeper.membership.hyparview.ActiveProtocol._
-import zio.keeper.membership.hyparview.InitialProtocol
-import zio.keeper.membership.hyparview.InitialProtocol._
-import zio.keeper.membership.hyparview.JoinReply
-import zio.keeper.membership.hyparview.NeighborReply
-import zio.keeper.membership.hyparview.NeighborReply._
-import zio.keeper.membership.swim.protocols.{ FailureDetection, Initial, Suspicion }
-import zio.keeper.membership.swim.protocols.FailureDetection.{ Ack, Nack, Ping, PingReq }
-import zio.keeper.membership.swim.protocols.Suspicion.{ Alive, Dead, Suspect }
+import zio.keeper.hyparview.ActiveProtocol
+import zio.keeper.hyparview.ActiveProtocol._
+import zio.keeper.hyparview.InitialProtocol
+import zio.keeper.hyparview.InitialProtocol._
+import zio.keeper.hyparview.JoinReply
+import zio.keeper.hyparview.NeighborReply
+import zio.keeper.hyparview.NeighborReply._
+import zio.keeper.swim.protocols.{ FailureDetection, Initial, Suspicion }
+import zio.keeper.swim.protocols.FailureDetection.{ Ack, Nack, Ping, PingReq }
+import zio.keeper.swim.protocols.Suspicion.{ Alive, Dead, Suspect }
 import zio.test._
 
 object ByteCodecSpec extends DefaultRunnableSpec {
