+++ keeper/src/test/scala/zio/keeper/gens.scala
@@ -3,13 +3,13 @@ package zio.keeper
 import java.util.UUID
 
 import zio.Chunk
-import zio.keeper.membership.hyparview.ActiveProtocol._
-import zio.keeper.membership.hyparview.InitialProtocol._
-import zio.keeper.membership.hyparview.NeighborReply._
-import zio.keeper.membership.hyparview._
-import zio.keeper.membership.swim.protocols.{ FailureDetection, Initial, Suspicion }
-import zio.keeper.membership.swim.protocols.FailureDetection.{ Ack, Nack, Ping, PingReq }
-import zio.keeper.membership.swim.protocols.Suspicion.{ Alive, Dead, Suspect }
+import zio.keeper.hyparview.ActiveProtocol._
+import zio.keeper.hyparview.InitialProtocol._
+import zio.keeper.hyparview.NeighborReply._
+import zio.keeper.hyparview._
+import zio.keeper.swim.protocols.{ FailureDetection, Initial, Suspicion }
+import zio.keeper.swim.protocols.FailureDetection.{ Ack, Nack, Ping, PingReq }
+import zio.keeper.swim.protocols.Suspicion.{ Alive, Dead, Suspect }
 import zio.random.Random
 import zio.test._
