+++ keeper/src/test/scala/zio/keeper/hyparview/ViewsSpec.scala
@@ -1,4 +1,4 @@
-package zio.keeper.membership.hyparview
+package zio.keeper.hyparview
 
 import zio._
 import zio.keeper.SendError.TransportFailed
